+++ databank/management/commands/sources/FTS_HPC.py
@@ -31,7 +31,7 @@ def prefetch():
         f"{d['Country #country+code'].upper()}-{d['Year #date+year']}": {
             "people_in_need": d["PIN #inneed"],
             "people_targeted": d["PT #targeted"],
-            "funding_total_usd": d["Funding #value+funding+total+usd"],
+            "funding_total_usd": d["Funding #value+required+total+usd"],
             "funding_required_usd": d["Requirements #value+funding+required+usd"],
         }
         for d in g_sheet_data